+++ Base/Graphics/Buffer/IndexBuffer.cpp
@@ -120,7 +120,7 @@ bool IndexBuffer::Initialize(uint64_t size, bool isSrv, bool isShortFormat)
         srvDesc.ViewDimension = D3D12_SRV_DIMENSION_BUFFER;
         srvDesc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
         srvDesc.Buffer.FirstElement = 0;
-        srvDesc.Buffer.NumElements = size / stride;
+        srvDesc.Buffer.NumElements = static_cast<UINT>(size / stride);
         srvDesc.Buffer.StructureByteStride = stride;
         srvDesc.Buffer.Flags = D3D12_BUFFER_SRV_FLAG_NONE;
