+++ Base/Graphics/Buffer/VertexBuffer.cpp
@@ -121,7 +121,7 @@ bool VertexBuffer::Initialize(uint64_t size, uint32_t stride, bool isSrv)
         srvDesc.ViewDimension = D3D12_SRV_DIMENSION_BUFFER;
         srvDesc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
         srvDesc.Buffer.FirstElement = 0;
-        srvDesc.Buffer.NumElements = size / stride;
+        srvDesc.Buffer.NumElements = static_cast<UINT>(size / stride);
         srvDesc.Buffer.StructureByteStride = stride;
         srvDesc.Buffer.Flags = D3D12_BUFFER_SRV_FLAG_NONE;
